requirements.txt
astroid==2.3.0
certifi==2019.9.11
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
psycopg2==2.8.3
pylint==2.4.1
pytz==2019.3
sqlparse==0.3.0
typed-ast==1.2.0
wincertstore==0.2
wrapt==1.11.2
